Move abstract class into the parent namespace. follow how standard in lazer did.
