+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Settings/IssueTableContainer.cs
@@ -15,7 +15,7 @@ using osu.Game.Graphics.Containers;
 using osu.Game.Graphics.Sprites;
 using osuTK.Graphics;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings
 {
     public abstract class IssueTableContainer : TableContainer
     {
