+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Settings/Language/LanguageAutoGenerateSubsection.cs
@@ -6,7 +6,6 @@
 using J2N.Collections.Generic;
 using osu.Framework.Graphics.UserInterface;
 using osu.Game.Rulesets.Karaoke.Edit.Configs.Generator.Language;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components.Markdown;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Language
