+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Settings/Language/LanguageSwitchSpecialActionSection.cs
@@ -4,7 +4,6 @@
 #nullable disable
 
 using osu.Framework.Allocation;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Language
