+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Settings/Notes/NoteClearSubsection.cs
@@ -7,7 +7,6 @@ using System.Collections.Generic;
 using osu.Framework.Allocation;
 using osu.Framework.Localisation;
 using osu.Game.Rulesets.Karaoke.Edit.ChangeHandlers.Notes;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components;
 using osu.Game.Rulesets.Karaoke.Objects;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Notes
