+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Settings/Reference/ReferenceLyricAutoGenerateSection.cs
@@ -4,7 +4,6 @@
 using osu.Framework.Graphics.UserInterface;
 using osu.Framework.Localisation;
 using osu.Game.Rulesets.Karaoke.Edit.Configs.Generator.ReferenceLyric;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components.Markdown;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Reference
