+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Settings/RubyRomaji/TextTagAutoGenerateSection.cs
@@ -5,7 +5,6 @@
 
 using J2N.Collections.Generic;
 using osu.Framework.Localisation;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components.Markdown;
 
 namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.RubyRomaji
