+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Settings/SpecialActionSection.cs
@@ -11,7 +11,7 @@ using osu.Game.Graphics.UserInterfaceV2;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes;
 
-namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components
+namespace osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings
 {
     public abstract class SpecialActionSection<TAction> : LyricEditorSection where TAction : struct, Enum
     {
