+++ osu.Game.Rulesets.Karaoke/Edit/Lyrics/Settings/TimeTags/TimeTagEditModeSection.cs
@@ -5,7 +5,6 @@ using System;
 using System.Collections.Generic;
 using osu.Game.Graphics;
 using osu.Game.Overlays;
-using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.Settings.Components.Markdown;
 using osu.Game.Rulesets.Karaoke.Edit.Lyrics.States.Modes;
 using osuTK.Graphics;
